Extract CppBackend::OutFile
I'll use this in a later commit to start compiling individual object
files.
